+++ mobile/android/android-components/components/browser/engine-gecko/src/main/java/mozilla/components/browser/engine/gecko/GeckoEngineSession.kt
@@ -37,6 +37,7 @@ import mozilla.components.concept.engine.history.HistoryItem
 import mozilla.components.concept.engine.history.HistoryTrackingDelegate
 import mozilla.components.concept.engine.manifest.WebAppManifest
 import mozilla.components.concept.engine.manifest.WebAppManifestParser
+import mozilla.components.concept.engine.pageextraction.ContentParams
 import mozilla.components.concept.engine.pageextraction.PageExtractionError
 import mozilla.components.concept.engine.pageextraction.PageMetadata
 import mozilla.components.concept.engine.request.RequestInterceptor
@@ -82,6 +83,7 @@ import org.mozilla.geckoview.GeckoSession.APP_LINK_LAUNCH_TYPE_WARM
 import org.mozilla.geckoview.GeckoSession.NavigationDelegate
 import org.mozilla.geckoview.GeckoSession.PermissionDelegate.ContentPermission
 import org.mozilla.geckoview.GeckoSessionSettings
+import org.mozilla.geckoview.PageExtractionController
 import org.mozilla.geckoview.WebRequestError
 import org.mozilla.geckoview.WebResponse
 import java.security.cert.X509Certificate
@@ -861,8 +863,15 @@ class GeckoEngineSession(
      * See [EngineSession.getPageContent]
      */
     @OptIn(ExperimentalGeckoViewApi::class)
-    override fun getPageContent(onResult: (String) -> Unit, onException: (Throwable) -> Unit) {
-        geckoSession.sessionPageExtractor.pageContent
+    override fun getPageContent(
+        options: ContentParams,
+        onResult: (String) -> Unit,
+        onException: (Throwable) -> Unit,
+    ) {
+        val geckoViewOptions = PageExtractionController.ContentParams(
+            options.removeBoilerplate,
+        )
+        geckoSession.sessionPageExtractor.getPageContent(geckoViewOptions)
             .then(
                 { content ->
                     if (content == null) {